Fix some sign compare warnings.

+++ src/ffx-fsr2-api/ffx_fsr2.cpp
@@ -179,7 +179,7 @@ static FfxErrorCode patchResourceBindings(FfxPipelineState* inoutPipeline)
 {
     for (uint32_t srvIndex = 0; srvIndex < inoutPipeline->srvCount; ++srvIndex)
     {
-        int32_t mapIndex = 0;
+        size_t mapIndex = 0;
         for (mapIndex = 0; mapIndex < _countof(srvResourceBindingTable); ++mapIndex)
         {
             if (0 == strcmp(srvResourceBindingTable[mapIndex].name, inoutPipeline->srvResourceBindings[srvIndex].name))
@@ -193,7 +193,7 @@ static FfxErrorCode patchResourceBindings(FfxPipelineState* inoutPipeline)
 
     for (uint32_t uavIndex = 0; uavIndex < inoutPipeline->uavCount; ++uavIndex)
     {
-        int32_t mapIndex = 0;
+        size_t mapIndex = 0;
         for (mapIndex = 0; mapIndex < _countof(uavResourceBindingTable); ++mapIndex)
         {
             if (0 == strcmp(uavResourceBindingTable[mapIndex].name, inoutPipeline->uavResourceBindings[uavIndex].name))
@@ -207,7 +207,7 @@ static FfxErrorCode patchResourceBindings(FfxPipelineState* inoutPipeline)
 
     for (uint32_t cbIndex = 0; cbIndex < inoutPipeline->constCount; ++cbIndex)
     {
-        int32_t mapIndex = 0;
+        size_t mapIndex = 0;
         for (mapIndex = 0; mapIndex < _countof(cbResourceBindingTable); ++mapIndex)
         {
             if (0 == strcmp(cbResourceBindingTable[mapIndex].name, inoutPipeline->cbResourceBindings[cbIndex].name))
